@@ -3,12 +3,12 @@ package com.app.alg.judix;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.navigation.NavigationView;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.TextView;
@@ -18,7 +18,6 @@ import com.app.alg.judix.telaMandado.ListaMandadoActivity;
 import com.app.alg.judix.telaNotificacaoMandado.ListaTodasNotificacoes;
 import com.app.alg.judix.telaOficial.ListaRelatorios;
 import com.app.alg.judix.telaOficial.OficialActivity;
-import com.app.alg.judix.telaOficial.RelatorioProdutividadeActivity;
 
 public class MenuActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener{
