@@ -14,17 +14,13 @@ import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Environment;
 import android.os.PowerManager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.ContextMenu;
-import android.view.LayoutInflater;
-import android.view.MenuInflater;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.Button;
 import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.android.volley.DefaultRetryPolicy;
@@ -36,16 +32,9 @@ import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
 import com.app.alg.judix.R;
 import com.app.alg.judix.model.ArquivoAnexoMandado;
-import com.app.alg.judix.model.AssinaturaMandado;
 import com.app.alg.judix.model.Mandado;
-import com.app.alg.judix.telaAssinaturasMandado.ItemAssinaturaMandadoAdapter;
-import com.app.alg.judix.telaAssinaturasMandado.ListaAssinaturasMandadoActivity;
-import com.app.alg.judix.telaMandado.ListaMandadoActivity;
-import com.app.alg.judix.telaMandado.MandadoActivity;
 import com.app.alg.judix.util.FilesHandler;
 import com.app.alg.judix.util.Funcoes;
-import com.github.barteksc.pdfviewer.PDFView;
-import com.github.barteksc.pdfviewer.listener.OnPageChangeListener;
 import com.google.gson.Gson;
 
 import org.json.JSONArray;
