@@ -7,7 +7,7 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.ContextMenu;
 import android.view.LayoutInflater;
@@ -31,7 +31,6 @@ import com.android.volley.toolbox.Volley;
 import com.app.alg.judix.R;
 import com.app.alg.judix.model.AssinaturaMandado;
 import com.app.alg.judix.model.Mandado;
-import com.app.alg.judix.telaAssinatura.Tela;
 import com.app.alg.judix.telaMandado.DocumentoDestinatarioMandado;
 import com.app.alg.judix.util.Funcoes;
 import com.google.gson.Gson;
