@@ -3,7 +3,7 @@ package com.app.alg.judix.telaNotificacaoMandado;
 import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -22,10 +22,8 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 import com.android.volley.toolbox.Volley;
 import com.app.alg.judix.R;
-import com.app.alg.judix.model.Certidao;
 import com.app.alg.judix.model.Mandado;
 import com.app.alg.judix.model.Notificacao;
-import com.app.alg.judix.telaOficial.AlterarDadosOficialActivity;
 import com.app.alg.judix.util.Funcoes;
 import com.app.alg.judix.util.Mask;
 import com.google.gson.Gson;
